requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
google-api-python-client==2.126.0
google-auth-oauthlib==1.2.0
apscheduler==3.10.4
requests==2.31.0
